# admin/static/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PLC4X Manager API",
    "description": "REST API for managing Apache PLC4X OPC-UA Server. Configure devices, read/write PLC tags, manage users, certificates, backups, and server lifecycle.",
    "version": "1.0.0",
    "contact": {
      "name": "Silvio Garbes",
      "url": "https://github.com/silviogarbes/plc4x-manager"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "servers": [
    {"url": "http://localhost:8080", "description": "Local (HTTP)"},
    {"url": "https://localhost:8443", "description": "Local (HTTPS)"}
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from POST /api/auth/login"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for machine-to-machine access (set API_KEY in .env)"
      }
    },
    "schemas": {
      "Device": {
        "type": "object",
        "required": ["name", "connectionString"],
        "properties": {
          "name": {"type": "string", "example": "ControlLogix1"},
          "connectionString": {"type": "string", "example": "logix://192.168.1.50"},
          "enabled": {"type": "boolean", "default": true},
          "tags": {"type": "array", "items": {"$ref": "#/components/schemas/Tag"}},
          "calculatedTags": {"type": "array", "items": {"$ref": "#/components/schemas/CalculatedTag"}}
        }
      },
      "Tag": {
        "type": "object",
        "required": ["alias", "address"],
        "properties": {
          "alias": {"type": "string", "example": "Temperature"},
          "address": {"type": "string", "example": "%DB1:0:REAL"}
        }
      },
      "CalculatedTag": {
        "type": "object",
        "required": ["alias", "formula"],
        "properties": {
          "alias": {"type": "string", "example": "TempFahrenheit"},
          "formula": {"type": "string", "example": "Temperature * 9 / 5 + 32"}
        }
      },
...
    }
  },
  "security": [{"BearerAuth": []}, {"ApiKeyAuth": []}],
...
  "paths": {
...
    "/api/devices": {
      "get": {
        "tags": ["Devices"],
        "summary": "List all devices",
        "responses": {"200": {"description": "Array of devices", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/Device"}}}}}}
      },
      "post": {
        "tags": ["Devices"],
        "summary": "Add a device",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Device"}}}},
        "responses": {"201": {"description": "Device added"}, "400": {"description": "Invalid input or tag address"}, "409": {"description": "Device already exists"}}
      }
    },
    "/api/devices/{name}": {
      "put": {
        "tags": ["Devices"],
        "summary": "Update a device",
        "parameters": [{"name": "name", "in": "path", "required": true, "schema": {"type": "string"}}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Device"}}}},
        "responses": {"200": {"description": "Device updated"}, "400": {"description": "Invalid tag address"}, "404": {"description": "Device not found"}}
      },
      "delete": {
        "tags": ["Devices"],
        "summary": "Delete a device",
        "parameters": [{"name": "name", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Device removed"}, "404": {"description": "Device not found"}}
      }
    },
...
  }
}